cli/metro-config: Move config loading to `@expo/metro-config`
