+++ packages/@expo/cli/src/start/server/metro/instantiateMetro.ts
@@ -9,13 +9,11 @@ import type MetroHmrServer from '@expo/metro/metro/HmrServer';
 import RevisionNotFoundError from '@expo/metro/metro/IncrementalBundler/RevisionNotFoundError';
 import type MetroServer from '@expo/metro/metro/Server';
 import formatBundlingError from '@expo/metro/metro/lib/formatBundlingError';
-import { mergeConfig, type ConfigT } from '@expo/metro/metro-config';
 import { Terminal } from '@expo/metro/metro-core';
 import type { createStableModuleIdFactory } from '@expo/metro-config';
-import { getDefaultConfig } from '@expo/metro-config';
+import { loadUserConfig } from '@expo/metro-config';
 import { patchTransformFileForPackedMaps } from '@expo/metro-config/build/serializer/packedMap';
 import { patchMetroSourceMapStringForPackedMaps } from '@expo/metro-config/build/serializer/sourceMap';
-import { resolveBabelrcName } from '@expo/metro-config/exports';
 import chalk from 'chalk';
 import type http from 'http';
 import path from 'path';
@@ -37,7 +35,6 @@ import { createCorsMiddleware } from '../middleware/CorsMiddleware';
 import { createJsInspectorMiddleware } from '../middleware/inspector/createJsInspectorMiddleware';
 import { prependMiddleware } from '../middleware/mutations';
 import { getPlatformBundlers } from '../platformBundlers';
-import { loadMetroConfigFileAsync } from './loadMetroConfigFile';
 
 // prettier-ignore
 export const event = events('metro', (t) => [
@@ -219,23 +216,15 @@ export async function loadMetroConfigAsync(
 
   const terminalReporter = new MetroTerminalReporter(serverRoot, terminal);
 
-  const defaultConfig = getDefaultConfig(projectRoot);
-  const resolvedConfig = await loadMetroConfigFileAsync({
+  let config = await loadUserConfig({
     projectRoot,
     serverRoot,
     // NOTE: Allow external tools to override the metro config. This is considered internal and unstable
     overrideConfigPath: env.EXPO_OVERRIDE_METRO_CONFIG ?? undefined,
   });
 
-  let config: ConfigT = resolvedConfig.isEmpty
-    ? defaultConfig
-    : await mergeConfig(defaultConfig, resolvedConfig.config);
-
-  // Set the watchfolders to include the projectRoot, as Metro assumes this
-  // Force-override the reporter
   config = {
     ...config,
-
     // See: `overrideConfigWithArguments` https://github.com/facebook/metro/blob/5059e26/packages/metro-config/src/loadConfig.js#L274-L339
     // Compare to `LoadOptions` type (disregard `reporter` as we don't expose this)
     resetCache: !!options.resetCache,
@@ -244,10 +233,7 @@ export async function loadMetroConfigAsync(
       ...config.server,
       port: options.port ?? config.server.port,
     },
-
-    watchFolders: !config.watchFolders.includes(config.projectRoot)
-      ? [config.projectRoot, ...config.watchFolders]
-      : config.watchFolders,
+    // Force-override the reporter
     reporter: {
       update(event) {
         terminalReporter.update(event);
@@ -258,23 +244,10 @@ export async function loadMetroConfigAsync(
     },
   };
 
-  // NOTE(@kitten): Pass a hint to the transformer on where to find the Babel config
-  asWritable(config.transformer).extendsBabelConfigPath =
-    config.transformer.enableBabelRCLookup !== false ? resolveBabelrcName(projectRoot) : undefined;
-
   // On-Demand Filesystem is enabled by default
   // TODO(@kitten): Add to config-types JSON schema
   const onDemandFilesystem = exp.experiments?.onDemandFilesystem ?? true;
   asWritable(config.resolver).unstable_onDemandFilesystem = onDemandFilesystem;
-
-  // NOTE(@kitten): `useWatchman` is currently enabled by default, but it also disables `forceNodeFilesystemAPI`.
-  // If we instead set it to the special value `null`, it gets enables but also bypasses the "native find" codepath,
-  // which is slower than just using the Node filesystem API
-  // See: https://github.com/facebook/metro/blob/b9c243f/packages/metro-file-map/src/index.js#L326
-  // See: https://github.com/facebook/metro/blob/b9c243f/packages/metro/src/node-haste/DependencyGraph/createFileMap.js#L109
-  if (config.resolver.useWatchman === true) {
-    asWritable(config.resolver).useWatchman = null as any;
-  }
 
   globalThis.__requireCycleIgnorePatterns = config.resolver?.requireCycleIgnorePatterns;
