+++ packages/@expo/metro-config/package.json
@@ -2,13 +2,13 @@
   "name": "@expo/metro-config",
   "version": "56.0.11",
   "description": "A Metro config for running React Native projects with the Metro bundler",
-  "main": "build/ExpoMetroConfig.js",
-  "types": "build/ExpoMetroConfig.d.ts",
+  "main": "build/index.js",
+  "types": "build/index.d.ts",
   "exports": {
     ".": {
-      "types": "./build/ExpoMetroConfig.d.ts",
-      "require": "./build/ExpoMetroConfig.js",
-      "default": "./build/ExpoMetroConfig.js"
+      "types": "./build/index.d.ts",
+      "require": "./build/index.js",
+      "default": "./build/index.js"
     },
     "./exports": {
       "types": "./build/exports.d.ts",
@@ -69,6 +69,7 @@
     "@expo/env": "workspace:~2.3.0",
     "@expo/json-file": "workspace:~10.2.0",
     "@expo/metro": "~56.0.0",
+    "@expo/require-utils": "workspace:^56.1.2",
     "@expo/spawn-async": "^1.8.0",
     "@jridgewell/gen-mapping": "^0.3.13",
     "@jridgewell/remapping": "^2.3.5",
